Move field reference and target reference under resolver-map module
